Moved findArticleFilePath & loadArticleData from ResourcesActions.js into ResourcesHelpers.js because they're helper functions not action creators
